# Optimise ra_log_segment:init