In particular the call the filename:absname/1 which isn't actually needed
was very slow. This should result in a ~40% speedup for the read_opt test case.

This also removes the `read_ahead` option which would never have been used given the strategy to read entries backwards (to be reviewed).


The `read_opt` test in ra_log_2_SUITE shows a substantial improvement in log initialisation time taken.

AFTER:
read took 0.389ms Reduction 1704
read init took 10.937ms Reduction 151330

BEFORE:

read took 0.42ms Reduction 3063
read init took 17.449ms Reduction 238108